refactor(stm): applied second round of comments
