+++ mithril-stm/src/protocol/aggregate_signature/mod.rs
@@ -20,15 +20,21 @@ use crate::{BaseFieldElement, StmResult, circuits::halo2_ivc::types::MessageHash
 
 #[cfg(feature = "future_snark")]
 #[derive(Clone, Debug)]
-pub struct GenesisMessagePreimage(pub Vec<u8>);
+pub struct MessagePreimage(pub Vec<u8>);
 
 #[cfg(feature = "future_snark")]
-impl TryInto<MessageHash> for &GenesisMessagePreimage {
+impl TryInto<MessageHash> for &MessagePreimage {
     type Error = anyhow::Error;
     fn try_into(self) -> StmResult<MessageHash> {
         let genesis_preimage_hash: [u8; 32] = Sha256::digest(self.0.clone()).into();
         let genesis_message_field_elem = BaseFieldElement::from_raw(&genesis_preimage_hash)?.0;
         Ok(MessageHash::from_field(genesis_message_field_elem))
+    }
+}
+
+impl MessagePreimage {
+    pub fn to_vec(&self) -> Vec<u8> {
+        self.0.clone()
     }
 }
